Create Set-MappedNetworkDrives.ps1
This script requires that the RSAT powershell tools are installed. If they aren't, it'll run the adjoining scripts to try and install them. It uses net-stat to create the drives, rather than New-PSDrive because I had inconsistent results with the persistence actually working. It does successfully test nested AD access.

The script needs to be run by a user with admin access (if you run it as the admin, it'll look for the wrong user permissions). Additionally, the drive letter is set based off of the 9th letter in the share folder's name. You may need to adjust this according to your environment or write specific if statements to accommodate how you do things.